requirements.txt
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==5.2.0
Markdown==3.4.1
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
zipp==3.11.0
